+++ document/remote/doc.go
@@ -1,0 +1,2 @@
+// Package remote renders documents through an HTTP render service.
+package remote